Added API key authentication

+++ server/api.js
@@ -5,19 +5,19 @@
 
 // Libraries
 const express = require('express');
-const app = express();
+const api = express();
 const mysql = require("mysql2/promise");
 const bodyParser = require('body-parser');
 const cors = require('cors');
+const fs = require('fs');
+var crypto = require("crypto");
 
 // Parse JSON and URL-encoded data
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(cors({
+api.use(bodyParser.json());
+api.use(bodyParser.urlencoded({ extended: true }));
+api.use(cors({
     origin: '*'
 }));
-
-const tables = ['name', 'time', 'date', 'score', 'coins'];
 
 // Database connection
 const pool = mysql.createPool({
@@ -30,6 +30,8 @@ const pool = mysql.createPool({
     ssl: { rejectUnauthorized: false }, // Necessary to connect to the database!!!
     connectionLimit: 10
 });
+
+const oegePassword = "pSmwQExG/1rux." // Password for creating a new API key
 
 /*========================*\
 |        Functions         |
@@ -46,8 +48,28 @@ function consoleLog(request, description) {
 \*========================*/
 
 // Get user data of all users
-app.get('/api/get/allusers', async (req, res) => {
+api.get('/api/get/allusers', async (req, res) => {
     consoleLog("GET", "all user data"); // Log the request
+
+    // Get the post data from the request:
+    let postData = JSON.stringify(req.body);
+    let parsedData = JSON.parse(postData);
+    let key = parsedData.key;
+
+    // Read the keys from the JSON file
+    let rawKeys = fs.readFileSync(`${__dirname}/keys.json`);
+    let keys = JSON.parse(rawKeys);
+
+    // Check if the API key is correct:
+    const validkeys = keys.key;
+    if (!validkeys.includes(key)) {
+        res.status(401); // HTTP Status 401: Unauthorized
+        const data = [{ message: 'Unauthorized' }];
+        res.json(data); // Send the response
+        return;
+    }
+
+    // Continue with the request:
     try {
         const conn = await pool.getConnection(); // Get a connection from the pool
         let sql = `SELECT * FROM \`userdata\`;`; // SQL query
@@ -67,7 +89,7 @@ app.get('/api/get/allusers', async (req, res) => {
 
 
 // Get user data of a specific user
-app.get('/api/get/user', async (req, res) => {
+api.get('/api/get/user', async (req, res) => {
 
     // Get the data from the request:
     let postData = JSON.stringify(req.body);
@@ -75,9 +97,24 @@ app.get('/api/get/user', async (req, res) => {
 
     // Get the data from the parsed data:
     let name = parsedData.name;
+    let key = parsedData.key;
+
+    // Read the keys from the JSON file
+    let rawKeys = fs.readFileSync(`${__dirname}/keys.json`);
+    let keys = JSON.parse(rawKeys);
 
     consoleLog("GET", `user data for user: ${name}`); // Log the request
 
+    // Check if the API key is correct:
+    const validkeys = keys.key;
+    if (!validkeys.includes(key)) {
+        res.status(401); // HTTP Status 401: Unauthorized
+        const data = [{ message: 'Unauthorized' }];
+        res.json(data); // Send the response
+        return;
+    }
+
+    // Continue with the request:
     try {
         const conn = await pool.getConnection(); // Get a connection from the pool
         let sql = `SELECT name, time, date, score, coins FROM \`userdata\` WHERE name = ?;`; // SQL query
@@ -98,9 +135,28 @@ app.get('/api/get/user', async (req, res) => {
 
 
 // Get the user with the highest score
-app.get('/api/get/mostscore', async (req, res) => {
+api.get('/api/get/mostscore', async (req, res) => {
     consoleLog("GET", "the highest score"); // Log the request
 
+    // Get the post data from the request:
+    let postData = JSON.stringify(req.body);
+    let parsedData = JSON.parse(postData);
+    let key = parsedData.key;
+
+    // Read the keys from the JSON file
+    let rawKeys = fs.readFileSync(`${__dirname}/keys.json`);
+    let keys = JSON.parse(rawKeys);
+
+    // Check if the API key is correct:
+    const validkeys = keys.key;
+    if (!validkeys.includes(key)) {
+        res.status(401); // HTTP Status 401: Unauthorized
+        const data = [{ message: 'Unauthorized' }];
+        res.json(data); // Send the response
+        return;
+    }
+
+    // Continue with the request:
     try {
         const conn = await pool.getConnection(); // Get a connection from the pool
         let sql = `SELECT MAX(score) FROM \`userdata\`;`; // SQL query
@@ -136,9 +192,28 @@ app.get('/api/get/mostscore', async (req, res) => {
 
 
 // Get the user with the most coins
-app.get('/api/get/mostcoins', async (req, res) => {
+api.get('/api/get/mostcoins', async (req, res) => {
     consoleLog("GET", "the most coins"); // Log the request
 
+    // Get the post data from the request:
+    let postData = JSON.stringify(req.body);
+    let parsedData = JSON.parse(postData);
+    let key = parsedData.key;
+
+    // Read the keys from the JSON file
+    let rawKeys = fs.readFileSync(`${__dirname}/keys.json`);
+    let keys = JSON.parse(rawKeys);
+
+    // Check if the API key is correct:
+    const validkeys = keys.key;
+    if (!validkeys.includes(key)) {
+        res.status(401); // HTTP Status 401: Unauthorized
+        const data = [{ message: 'Unauthorized' }];
+        res.json(data); // Send the response
+        return;
+    }
+
+    // Continue with the request:
     try {
         const conn = await pool.getConnection(); // Get a connection from the pool
         let sql = `SELECT MAX(coins) FROM \`userdata\`;`; // SQL query
@@ -168,14 +243,34 @@ app.get('/api/get/mostcoins', async (req, res) => {
         result = [{ message: 'Error' }, { error: err }];
     }
     res.json(result); // Send the response, only send the data, not the metadata
+    return;
 });
 
 
 
 // Get the top 10 users with the highest score
-app.get('/api/get/leaderboard/score', async (req, res) => {
+api.get('/api/get/leaderboard/score', async (req, res) => {
     consoleLog("GET", "leaderboard score"); // Log the request
 
+    // Get the post data from the request:
+    let postData = JSON.stringify(req.body);
+    let parsedData = JSON.parse(postData);
+    let key = parsedData.key;
+
+    // Read the keys from the JSON file
+    let rawKeys = fs.readFileSync(`${__dirname}/keys.json`);
+    let keys = JSON.parse(rawKeys);
+
+    // Check if the API key is correct:
+    const validkeys = keys.key;
+    if (!validkeys.includes(key)) {
+        res.status(401); // HTTP Status 401: Unauthorized
+        const data = [{ message: 'Unauthorized' }];
+        res.json(data); // Send the response
+        return;
+    }
+
+    // Continue with the request:
     try {
         const conn = await pool.getConnection(); // Get a connection from the pool
         let sql = `SELECT * FROM \`userdata\` ORDER BY \`userdata\`.\`score\` DESC LIMIT 10`; // SQL query
@@ -188,74 +283,34 @@ app.get('/api/get/leaderboard/score', async (req, res) => {
         result = [{ message: 'Error' }, { error: err }];
     }
     res.json(result[0]); // Send the response, only send the data, not the metadata
-
+    return;
 });
 
-
-const maxResults = 100;
-
-// Get request, for users to request specific kinds of data
-// One can provide a body containing request parameters, such as how many objects one wants to retrieve
-// A JSON request body will look like the following, if all properties are present:
-// {
-//  "results": 100,                         How many results will be returned
-//  "orderBy": "coins",                     Which table to filter by
-//  "tables": ["name", "coins", "score"],   Which table to select, can also be '*' to be any
-// }
-app.post('/api/get', (req, res) => {
-    console.log("Retrieved post request");
-    console.log(req.body);
-
-    let resultCount = maxResults;
-    let ordered = null;
-
-    // If the body contains a 'results' parameter of type 'number', the server
-    // will return that many results from the database.
-    if (typeof req.body.results === 'number')
-        resultCount = req.body.results;
-
-    // If the provided JSON body contains an 'orderBy' tag of type 'string', and the string
-    // is a valid table, as defined above, it will order the query by that table.
-    if (typeof req.body.orderBy === 'string' && tables.includes(req.body.orderBy))
-        ordered = req.body.orderBy;
-
-    if (req.body.tables === "*")
-        req.body.tables = tables;
-
-    if (Array.isArray(req.body.tables)) {
-        // Check whether the content of 'req.body.tables' contains acceptable tables
-        let receivedTables = req.body.tables.filter((key) => tables.includes(key));
-
-        // Check if there are any accepted tables
-        if (receivedTables.length > 0) {
-
-            // SQL query for looking up things from the database.
-            // When the user provides tables to look up, it will return the table
-            // If the user provides a limit, it will add "LIMIT x" to the query
-            // If the user provides 'orderBy' tag, and then provides an appropriate table, it will sort by that table
-            let sqlQuery =
-                `SELECT ${receivedTables.toString()} FROM userdata${ordered != null ? ` ORDER BY ${ordered} DESC` : ""} ${resultCount < 0 ? "" : `LIMIT ${resultCount}`}`;
-
-            (async () => {
-                // Open connection and make the produced query
-                let connection = await pool.getConnection();
-                await connection.query(sqlQuery)
-                    .then(result => res.json(result[0]))
-                    .catch((e) => console.log("Error occurred"))
-                    .finally(() => res.status(400));
-                connection.release();
-            })();
-
-            console.log(sqlQuery);
-        }
-    }
-});
 
 
 // Get the top 10 users with the most coins
-app.get('/api/get/leaderboard/coins', async (req, res) => {
+api.get('/api/get/leaderboard/coins', async (req, res) => {
     consoleLog("GET", "leaderboard coins"); // Log the request
 
+    // Get the post data from the request:
+    let postData = JSON.stringify(req.body);
+    let parsedData = JSON.parse(postData);
+    let key = parsedData.key;
+
+    // Read the keys from the JSON file
+    let rawKeys = fs.readFileSync(`${__dirname}/keys.json`);
+    let keys = JSON.parse(rawKeys);
+
+    // Check if the API key is correct:
+    const validkeys = keys.key;
+    if (!validkeys.includes(key)) {
+        res.status(401); // HTTP Status 401: Unauthorized
+        const data = [{ message: 'Unauthorized' }];
+        res.json(data); // Send the response
+        return;
+    }
+
+    // Continue with the request:
     try {
         const conn = await pool.getConnection(); // Get a connection from the pool
         let sql = `SELECT * FROM \`userdata\` ORDER BY \`userdata\`.\`coins\` DESC LIMIT 10`; // SQL query
@@ -268,12 +323,13 @@ app.get('/api/get/leaderboard/coins', async (req, res) => {
         result = [{ message: 'Error' }, { error: err }];
     }
     res.json(result[0]); // Send the response, only send the data, not the metadata
+    return;
 });
 
 
 
 // Insert user data
-app.post('/api/post/insert', async (req, res) => {
+api.post('/api/post/insert', async (req, res) => {
 
     // Get the data from the request:
     let postData = JSON.stringify(req.body);
@@ -282,14 +338,28 @@ app.post('/api/post/insert', async (req, res) => {
 
     // Get the data from the parsed data:
     let name = parsedData.name;
-    let _date = new Date();
-    let time = _date.toLocaleTimeString();
-    let date = _date.toLocaleDateString();
+    let time = getTime();
+    let date = getDateUS();
     let score = parsedData.score;
     let coins = parsedData.coins;
+    let key = parsedData.key;
 
     consoleLog("POST", `inserting user data for: ${name}`); // Log the request
 
+    // Read the keys from the JSON file
+    let rawKeys = fs.readFileSync(`${__dirname}/keys.json`);
+    let keys = JSON.parse(rawKeys);
+
+    // Check if the API key is correct:
+    const validkeys = keys.key;
+    if (!validkeys.includes(key)) {
+        res.status(401); // HTTP Status 401: Unauthorized
+        data = [{ message: 'Unauthorized' }];
+        res.json(data); // Send the response
+        return;
+    }
+
+    // Continue with the request:
     try {
         // Make a connection to the database:
         const conn = await pool.getConnection(); // Get a connection from the pool
@@ -312,7 +382,7 @@ app.post('/api/post/insert', async (req, res) => {
 
 
 // Delete user data of a specific user
-app.delete('/api/delete/user', async (req, res) => {
+api.delete('/api/delete/user', async (req, res) => {
 
     // Get the data from the request:
     let postData = JSON.stringify(req.body);
@@ -320,9 +390,24 @@ app.delete('/api/delete/user', async (req, res) => {
 
     // Get the data from the parsed data:
     let name = parsedData.name;
+    let key = parsedData.key;
 
     consoleLog("DELETE", `deleting user data for: ${name}`); // Log the request
 
+    // Read the keys from the JSON file
+    let rawKeys = fs.readFileSync(`${__dirname}/keys.json`);
+    let keys = JSON.parse(rawKeys);
+
+    // Check if the API key is correct:
+    const validkeys = keys.key;
+    if (!validkeys.includes(key)) {
+        res.status(401); // HTTP Status 401: Unauthorized
+        const data = [{ message: 'Unauthorized' }];
+        res.json(data); // Send the response
+        return;
+    }
+
+    // Continue with the request:
     try {
         const conn = await pool.getConnection(); // Get a connection from the pool
         let sql = `DELETE FROM \`userdata\` WHERE name = ?;`; // SQL query
@@ -338,13 +423,33 @@ app.delete('/api/delete/user', async (req, res) => {
         result = [{ message: 'Error' }, { error: err }];
     }
     res.json(result); // Send the response
+    return;
 });
 
 
 
 // Test the API
-app.get('/api/test', (req, res) => {
+api.get('/api/test', (req, res) => {
     consoleLog("GET", "test"); // Log the request
+    // Get the data from the request:
+    let postData = JSON.stringify(req.body);
+    let parsedData = JSON.parse(postData);
+    let key = parsedData.key;
+
+    // Read the keys from the JSON file
+    let rawKeys = fs.readFileSync(`${__dirname}/keys.json`);
+    let keys = JSON.parse(rawKeys);
+
+    // Check if the API key is correct:
+    const validkeys = keys.key;
+    if (!validkeys.includes(key)) {
+        res.status(401); // HTTP Status 401: Unauthorized
+        const data = [{ message: 'Unauthorized' }];
+        res.json(data); // Send the response
+        return;
+    }
+
+    // Authorized, continue with the request:
     const data = [{ message: 'API Success' }];
     res.status(200); // HTTP Status 200: OK
     res.json(data); // Send the response
@@ -352,8 +457,45 @@ app.get('/api/test', (req, res) => {
 
 
 
+// Create a new API key
+api.get ('/api/createkey', (req, res) => {
+
+    // Get the data from the request:
+    let postData = JSON.stringify(req.body);
+    let parsedData = JSON.parse(postData);
+    let password = parsedData.password;
+
+    // Check if the password is correct:
+    if (password != oegePassword) {
+        res.status(401); // HTTP Status 401: Unauthorized
+        const data = [{ message: 'Unauthorized' }];
+        res.json(data); // Send the response
+        return;
+    }
+
+    // Authorized, continue with the request:
+    // Generate a new API key:
+    let key = crypto.randomBytes(12).toString('hex');
+    let rawKeys = fs.readFileSync(`${__dirname}/keys.json`);
+    
+    // Insert the new key in the JSON file:
+    let keys = JSON.parse(rawKeys);
+    keys.key.push(key);
+    let newKeys = JSON.stringify(keys);
+    fs.writeFileSync(`${__dirname}/keys.json`, newKeys);
+
+    // Send the response:
+    res.status(201); // HTTP Status 201: Created
+    const data = [{ message: 'API key created', key: key}];
+    res.json(data); // Send the response
+    return;
+});
+    
+
+
+
 // Default (wrong URL(GET)) 
-app.get('/*', (req, res) => {
+api.get('/*', (req, res) => {
     consoleLog("GET", "wrong URL"); // Log the request
     const data = [{ message: 'Wrong URL' }];
     res.status(404); // HTTP Status 404: Not Found
@@ -363,22 +505,26 @@ app.get('/*', (req, res) => {
 
 
 // Default (wrong URL(POST))
-app.post('/*', (req, res) => {
+api.post('/*', (req, res) => {
     consoleLog("POST", "wrong URL"); // Log the request
     const data = [{ message: 'Wrong URL' }];
     res.status(404); // HTTP Status 404: Not Found
     res.json(data); // Send the response
 });
 
+
+
 // Default (wrong URL(DELETE))
-app.delete('/*', (req, res) => {
+api.delete('/*', (req, res) => {
     consoleLog("DELETE", "wrong URL"); // Log the request
     const data = [{ message: 'Wrong URL' }];
     res.status(404); // HTTP Status 404: Not Found
     res.json(data); // Send the response
 });
 
+
+
 // Run the API on port 8080
-app.listen(8080, () => {
+api.listen(8080, () => {
     consoleLog("LISTEN", "running API on port 8080"); // Log the request
-});
+});
